Fix validation for invalid Base58 addresses
If the address is invalid we can't derive the network from it so
we must return an empty string to avoid hitting an exception
later on.

e.g. TypeError: ("'network' must be <class 'str'> (got None that is a <class 'NoneType'>).",
To return an empty network string for when the address is invalid. Seems almost all bug reports are about this same issue.